Move new tests into already existing test classes
